add new features

+++ app/Http/Controllers/Backend/ConversationLogController.php
@@ -14,7 +14,16 @@ class ConversationLogController extends Controller
 {
     public function index()
     {
-        return view('backend.Conversation_log.index');
+        if (auth()->check()) {
+            if (auth()->user()->can('conversation')) {
+                return view('backend.Conversation_log.index');
+            } else {
+                auth()->logout(); // Log out the user
+                return redirect()->route('login')->with('error', 'You do not have permission to view and have been logged out.');
+            }
+        } else {
+            return redirect()->back()->with('error', 'You need to login first.');
+        }
     }
 
 
@@ -29,17 +38,26 @@ class ConversationLogController extends Controller
                 ->addColumn('project', function ($row) {
                     return $row->project->name;
                 })
+                ->addColumn('date', function ($row) {
+                    return $row->date ? $row->date : 'N/A';
+                })
                 ->addColumn('action', function ($row) {
+                    $editBtn='';
+                    $deleteBtn='';
+
                     $editUrl = route('conversation-logs.edit', $row->id);
                     $deleteUrl = route('conversation-logs.distroy', $row->id);
                     $csrfToken = csrf_field();
                     $method = method_field('DELETE');
 
-
+                    if (auth()->user()->can('conversation-edit')) {
                     $editBtn = '<a href="' . $editUrl . '" class="edit btn btn-sm btn-success me-2 rounded" style="padding:8px;"><span>' .
                         '<svg viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg" style="width: 20px; height: 20px;">' .
                         '<g id="SVGRepo_iconCarrier"> <path fill-rule="evenodd" clip-rule="evenodd" d="M21.1213 2.70705C19.9497 1.53548 18.0503 1.53547 16.8787 2.70705L15.1989 4.38685L7.29289 12.2928C7.16473 12.421 7.07382 12.5816 7.02986 12.7574L6.02986 16.7574C5.94466 17.0982 6.04451 17.4587 6.29289 17.707C6.54127 17.9554 6.90176 18.0553 7.24254 17.9701L11.2425 16.9701C11.4184 16.9261 11.5789 16.8352 11.7071 16.707L19.5556 8.85857L21.2929 7.12126C22.4645 5.94969 22.4645 4.05019 21.2929 2.87862L21.1213 2.70705ZM18.2929 4.12126C18.6834 3.73074 19.3166 3.73074 19.7071 4.12126L19.8787 4.29283C20.2692 4.68336 20.2692 5.31653 19.8787 5.70705L18.8622 6.72357L17.3068 5.10738L18.2929 4.12126ZM15.8923 6.52185L17.4477 8.13804L10.4888 15.097L8.37437 15.6256L8.90296 13.5112L15.8923 6.52185ZM4 7.99994C4 7.44766 4.44772 6.99994 5 6.99994H10C10.5523 6.99994 11 6.55223 11 5.99994C11 5.44766 10.5523 4.99994 10 4.99994H5C3.34315 4.99994 2 6.34309 2 7.99994V18.9999C2 20.6568 3.34315 21.9999 5 21.9999H16C17.6569 21.9999 19 20.6568 19 18.9999V13.9999C19 13.4477 18.5523 12.9999 18 12.9999C17.4477 12.9999 17 13.4477 17 13.9999V18.9999C17 19.5522 16.5523 19.9999 16 19.9999H5C4.44772 19.9999 4 19.5522 4 18.9999V7.99994Z" fill="#ffffff"></path> </g></svg>' .
                         '</span></a>';
+                    }
+
+                    if (auth()->user()->can('conversation-delete')) {
                     $deleteBtn = '<form action="' . $deleteUrl . '" method="POST">
                     ' . $csrfToken . '
                     ' . $method . '
@@ -49,20 +67,21 @@ class ConversationLogController extends Controller
                             </svg>
                             </button>
                 </form>';
+                    }
                     return $editBtn . ' ' . $deleteBtn;
                 })
                 ->rawColumns(['action'])
                 ->make(true);
         }
     }
-   
+
 
     public function create()
     {
         $customers = Customer::all();
         return view('backend.Conversation_log.create', compact('customers'));
     }
-    
+
     public function getAssignedProject($id)
     {
         $customer = Customer::find($id);
@@ -77,46 +96,70 @@ class ConversationLogController extends Controller
 
     public function store(Request $request)
     {
-
         $request->validate([
             'customer_id' => 'required|exists:customers,id',
             'project_id' => 'required|exists:projects,id',
-            'note' => 'nullable|string'
+            'note' => 'nullable|string',
+            'date' => 'nullable|date|date_format:Y-m-d',
         ]);
 
         ConversationLog::create([
             'customer_id' => $request->customer_id,
             'project_id' => $request->project_id,
-            'note' => $request->note
+            'note' => $request->note,
+            'date' => $request->date
         ]);
 
         return redirect()->back()->with('success', 'Data inserted successfull');
     }
 
-    public function edit($id){
-        $customerLog= ConversationLog::findOrFail($id);
-        $customers=Customer::all();
-        return view('backend.Conversation_log.edit',compact('customerLog','customers'));
+    public function edit($id)
+    {
+        if (auth()->check()) {
+            if (auth()->user()->can('conversation-edit')) {
+                $customerLog = ConversationLog::findOrFail($id);
+                $customers = Customer::all();
+                return view('backend.Conversation_log.edit', compact('customerLog', 'customers'));
+            } else {
+                auth()->logout(); // Log out the user
+                return redirect()->route('login')->with('error', 'You do not have permission to edit and have been logged out.');
+            }
+        } else {
+            return redirect()->back()->with('error', 'You need to login first.');
+        }
     }
 
-    public function update(Request $request,$id){
+    public function update(Request $request, $id)
+    {
         $request->validate([
             'customer_id' => 'required|exists:customers,id',
             'project_id' => 'required|exists:projects,id',
-            'note' => 'nullable|string'
+            'note' => 'nullable|string',
+            'date' => 'nullable|date|date_format:Y-m-d',
         ]);
 
-        $data=ConversationLog::findOrFail($id);
+        $data = ConversationLog::findOrFail($id);
         $data->customer_id = $request->customer_id;
         $data->project_id = $request->project_id;
         $data->note = $request->note;
+        $data->date = $request->date;
         $data->save();
         return redirect()->route('conversation-logs')->with('success', 'Data updated successfull');
     }
 
-    public function distroy($id){
-        $data=ConversationLog::findOrFail($id);
-        $data->delete();
-        return redirect()->route('conversation-logs')->with('success', 'Data deleted successfull');
+    public function distroy($id)
+    {
+        if (auth()->check()) {
+            if (auth()->user()->can('conversation-delete')) {
+                $data = ConversationLog::findOrFail($id);
+                $data->delete();
+                return redirect()->route('conversation-logs')->with('success', 'Data deleted successfull');
+            } else {
+                auth()->logout(); // Log out the user
+                return redirect()->route('login')->with('error', 'You do not have permission to delete and have been logged out.');
+            }
+        } else {
+            return redirect()->back()->with('error', 'You need to login first.');
+        }
     }
 }
